requirements.txt
Django>=2.1,<3.0
djangorestframework
django_rest_swagger
